+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 * Increased testing coverage
 * Some examples include links to Jupyter notebooks
 * Errors in documentation building are now tested in Travis
+* Improved batch processor
 
 
 0.6 (2018-02-05)